app[REFACT](VM) moved stripe fetch payment in VM
